# Replay WAL concurrently without blocking