@colega 2022-07-05
This is definitely much cleaner now! My comments are optional and opinionated nitpicks.