@damnever 2022-07-05
@jesusvazquez I have run the [benchmark test](https://github.com/prometheus/prometheus/blob/4b9f248e8569fc5163841c2e21b8fa5d3a61b311/tsdb/head_test.go#L134-L263) against the `main` branch, and interestingly there is no significant difference on my old mac. I may look into this later.