@damnever 2022-07-06
The benchmark result looks better if I modify test like this:
```patch
@@ -185,9 +185,13 @@ func BenchmarkLoadWAL(b *testing.B) {
 
 					// Write series.
 					refSeries := make([]record.RefSeries, 0, c.seriesPerBatch)
+					refSamples := make([]record.RefSample, 0, c.seriesPerBatch)
 					for k := 0; k < c.batches; k++ {
 						refSeries = refSeries[:0]
 						for i := k * c.seriesPerBatch; i < (k+1)*c.seriesPerBatch; i++ {
+							if i%2 == 0 {
+								continue
+							}
 							lbls := make(map[string]string, labelsPerSeries)
 							lbls[defaultLabelName] = strconv.Itoa(i)
 							for j := 1; len(lbls) < labelsPerSeries; j++ {
@@ -196,22 +200,48 @@ func BenchmarkLoadWAL(b *testing.B) {
 							refSeries = append(refSeries, record.RefSeries{Ref: chunks.HeadSeriesRef(i) * 101, Labels: labels.FromMap(lbls)})
 						}
 						populateTestWAL(b, w, []interface{}{refSeries})
-					}
+						// Write samples.
+						for i := k * c.seriesPerBatch; i < (k+1)*c.seriesPerBatch; i++ {
+							if i%2 == 0 {
+								continue
+							}
+							refSamples = refSamples[:0]
+							for j := 0; j < c.samplesPerSeries; j++ {
+								refSamples = append(refSamples, record.RefSample{
+									Ref: chunks.HeadSeriesRef(i) * 101,
+									T:   int64(j) * 10,
+									V:   float64(j) * 100,
+								})
 
-					// Write samples.
-					refSamples := make([]record.RefSample, 0, c.seriesPerBatch)
-					for i := 0; i < c.samplesPerSeries; i++ {
-						for j := 0; j < c.batches; j++ {
+							}
+							populateTestWAL(b, w, []interface{}{refSamples})
+						}
+
+						refSeries = refSeries[:0]
+						for i := k * c.seriesPerBatch; i < (k+1)*c.seriesPerBatch; i++ {
+							if i%2 != 0 {
+								continue
+							}
+							lbls := make(map[string]string, labelsPerSeries)
+							lbls[defaultLabelName] = strconv.Itoa(i)
+							for j := 1; len(lbls) < labelsPerSeries; j++ {
+								lbls[defaultLabelName+strconv.Itoa(j)] = defaultLabelValue + strconv.Itoa(j)
+							}
+							refSeries = append(refSeries, record.RefSeries{Ref: chunks.HeadSeriesRef(i) * 101, Labels: labels.FromMap(lbls)})
+
+							// Write samples.
 							refSamples = refSamples[:0]
-							for k := j * c.seriesPerBatch; k < (j+1)*c.seriesPerBatch; k++ {
+							for j := 0; j < c.samplesPerSeries; j++ {
 								refSamples = append(refSamples, record.RefSample{
-									Ref: chunks.HeadSeriesRef(k) * 101,
-									T:   int64(i) * 10,
-									V:   float64(i) * 100,
+									Ref: chunks.HeadSeriesRef(i) * 101,
+									T:   int64(j) * 10,
+									V:   float64(j) * 100,
 								})
+
 							}
 							populateTestWAL(b, w, []interface{}{refSamples})
 						}
+						populateTestWAL(b, w, []interface{}{refSeries})
 					}
 
 					// Write mmapped chunks.
```

```
goos: darwin
goarch: amd64
name                                                                                                      old time/op    new time/op    delta
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0-8          162ms ± 2%     147ms ± 4%   -9.46%  (p=0.000 n=10+9)
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0-8         166ms ± 4%     154ms ± 8%   -7.47%  (p=0.001 n=9+10)
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0-8         170ms ± 3%     156ms ± 4%   -8.07%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0-8        209ms ± 2%     191ms ± 3%   -8.49%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0-8          580ms ± 5%     528ms ± 4%   -8.88%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0-8          629ms ± 6%     562ms ± 5%  -10.75%  (p=0.000 n=9+9)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0-8          159ms ± 4%     151ms ± 6%   -4.78%  (p=0.002 n=10+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0-8          162ms ± 3%     151ms ± 4%   -6.50%  (p=0.000 n=9+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0-8          165ms ± 4%     155ms ± 5%   -6.08%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0-8         191ms ± 3%     180ms ± 4%   -6.04%  (p=0.000 n=10+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800-8      1.23s ± 5%     1.07s ± 3%  -12.34%  (p=0.000 n=10+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800-8      1.28s ± 5%     1.14s ± 3%  -11.13%  (p=0.000 n=10+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800-8      1.34s ±12%     1.16s ± 3%  -13.48%  (p=0.000 n=10+9)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800-8     1.67s ± 4%     1.49s ± 3%  -10.98%  (p=0.000 n=9+9)

name                                                                                                      old alloc/op   new alloc/op   delta
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0-8          328MB ± 3%     284MB ± 5%  -13.64%  (p=0.000 n=10+9)
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0-8         335MB ± 2%     287MB ± 8%  -14.40%  (p=0.000 n=9+9)
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0-8         336MB ± 6%     294MB ± 5%  -12.38%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0-8        361MB ± 6%     301MB ± 3%  -16.79%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0-8          183MB ± 2%     164MB ± 1%  -10.32%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0-8          224MB ± 2%     203MB ± 2%   -9.39%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0-8         79.8MB ± 8%    44.3MB ± 2%  -44.45%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0-8         82.2MB ± 5%    47.8MB ± 4%  -41.83%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0-8         84.6MB ± 3%    50.4MB ± 4%  -40.46%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0-8        99.1MB ± 2%    64.7MB ± 3%  -34.70%  (p=0.000 n=9+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800-8      798MB ±20%     290MB ± 7%  -63.63%  (p=0.000 n=10+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800-8      790MB ±12%     302MB ±13%  -61.82%  (p=0.000 n=10+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800-8      812MB ±20%     343MB ± 7%  -57.79%  (p=0.000 n=9+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800-8     955MB ± 5%     476MB ± 4%  -50.19%  (p=0.000 n=8+9)

name                                                                                                      old allocs/op  new allocs/op  delta
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0-8           256k ± 0%      313k ± 2%  +22.30%  (p=0.000 n=10+9)
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0-8          401k ± 0%      464k ± 0%  +15.83%  (p=0.000 n=9+7)
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0-8          547k ± 0%      609k ± 0%  +11.49%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0-8        1.71M ± 0%     1.75M ± 0%   +2.65%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0-8          2.37M ± 0%     2.29M ± 0%   -3.60%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0-8          3.17M ± 0%     3.09M ± 0%   -2.72%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0-8           380k ± 1%      357k ± 0%   -6.06%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0-8           460k ± 0%      437k ± 0%   -4.92%  (p=0.000 n=9+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0-8           580k ± 1%      557k ± 0%   -3.99%  (p=0.000 n=10+10)
LoadWALX/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0-8         1.34M ± 0%     1.32M ± 0%   -1.69%  (p=0.000 n=9+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800-8      3.27M ± 2%     2.96M ± 0%   -9.39%  (p=0.000 n=10+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800-8      4.06M ± 1%     3.76M ± 0%   -7.33%  (p=0.000 n=10+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800-8      5.27M ± 2%     4.97M ± 0%   -5.71%  (p=0.000 n=10+10)
LoadWALX/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800-8     12.9M ± 0%     12.6M ± 0%   -2.23%  (p=0.000 n=8+9)
```